+++ src/dbt_score/__init__.py
@@ -1,10 +1,11 @@
 """Init dbt_score package."""
 
-from dbt_score.models import Model, Seed, Snapshot, Source
+from dbt_score.models import Exposure, Model, Seed, Snapshot, Source
 from dbt_score.rule import Rule, RuleViolation, Severity, rule
 from dbt_score.rule_filter import RuleFilter, rule_filter
 
 __all__ = [
+    "Exposure",
     "Model",
     "Source",
     "Snapshot",
@@ -13,6 +14,6 @@ __all__ = [
     "Rule",
     "RuleViolation",
     "Severity",
-    "rule_filter",
     "rule",
+    "rule_filter",
 ]