+++ src/dbt_score/dbt_utils.py
@@ -75,6 +75,7 @@ def dbt_ls(select: Iterable[str] | None) -> Iterable[str]:
         "model",
         "source",
         "snapshot",
+        "exposure",
         "seed",
         "--output",
         "name",