+++ src/dbt_score/formatters/human_readable_formatter.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from dbt_score.evaluation import EvaluableResultsType
 from dbt_score.formatters import Formatter
-from dbt_score.models import Evaluable, Model, Seed, Snapshot, Source
+from dbt_score.models import Evaluable, Exposure, Model, Seed, Snapshot, Source
 from dbt_score.rule import RuleViolation
 from dbt_score.scoring import Score
 
@@ -36,6 +36,8 @@ class HumanReadableFormatter(Formatter):
             case Source():
                 return evaluable.selector_name
             case Snapshot():
+                return evaluable.name
+            case Exposure():
                 return evaluable.name
             case Seed():
                 return evaluable.name