+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 ## [Unreleased]
 
 - Support linting of exposures (#112)
+- Fix retrieval of data_tests when using quoted columns in BigQuery (#122)
 
 ## [0.12.0] - 2025-05-06